added a check for the settings.py file to the install.sh script. shifted the creation of the settings symlink before calling the install script

+++ install.sh
@@ -4,6 +4,9 @@ if [ ! -d venv ]; then
     virtualenv --python=`which python2` venv
 fi
 source venv/bin/activate
-
+if [ ! -f src/core/settings.py ]; then
+    print "no settings.py found. quitting while I'm ahead."
+    exit 1
+fi
 pip install -r requirements.txt
 python src/manage.py migrate
